Change types according to TypeLib

+++ Eco.Java2ACOM/SharedFiles/Eco/InterfaceBus/IEcoInterfaceBus1FileExt.java
@@ -9,10 +9,10 @@ import Eco.Core.UGUID;
 public interface IEcoInterfaceBus1FileExt extends IEcoUnknown {
     UGUID IID = new UGUID((byte) 0x01, (byte) 0x10, new byte[]{0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, (byte) 0xA0, 0x02, 0x00, 0x00, 0x01, 0x01});
 
-    int set_Manager(UGUID rcid);
+    short set_Manager(UGUID rcid);
     UGUID get_Manager();
-    int set_SearchPath(String path);
+    short set_SearchPath(String path);
     String get_SearchPath();
-    int RegisterComponent(UGUID rcid, String filename);
-    int QueryComponent(String filename, UGUID rcid, IEcoUnknown iUnkOuter, UGUID riid, IEcoUnknownPtr pIUnk);
+    short RegisterComponent(UGUID rcid, String filename);
+    short QueryComponent(String filename, UGUID rcid, IEcoUnknown iUnkOuter, UGUID riid, IEcoUnknownPtr pIUnk);
 }
