+++ Eco.Java2ACOM/SourceFiles/java/Eco/System/EcoMain.java
@@ -7,7 +7,7 @@ import java.util.ServiceLoader;
 class EcoMain {
     public static void main(String[] args) {
         IEcoSystem1Native iSys = new IEcoSystem1Native();
-        int result = CEcoSystem.createCEcoSystem(args, iSys);
+        short result = CEcoSystem.createCEcoSystem(args, iSys);
         if (result != 0) {
             System.exit(result);
         }
